requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.23
Werkzeug==1.0.1
Flask-Mail
Python-dotenv
Twilio